requirements.txt
-i https://pypi.org/simple
alembic==1.1.0
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
click==7.0
clickclick==1.2.2
connexion==2.3.0
entrypoints==0.3
flake8==3.7.7
flask-marshmallow==0.10.1
flask-migrate==2.5.2
flask-script==2.0.6
flask-sqlalchemy==2.4.0
flask==1.1.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
jinja2==2.10.1
jsonschema==2.6.0
mako==1.1.0
markupsafe==1.1.1
marshmallow-sqlalchemy==0.17.2
marshmallow==3.0.1
mccabe==0.6.1
openapi-spec-validator==0.2.8
psycopg2
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
python-editor==1.0.4
pyyaml==5.1.2
requests==2.22.0
six==1.12.0
sqlalchemy-utils==0.34.2
sqlalchemy==1.3.8
swagger-ui-bundle==0.0.5
toml==0.10.0
urllib3==1.25.3
werkzeug==0.15.5

